Move Local and Remote FeedItem types to their own files
