+++ EssentialFeed/EssentialFeed/Feed Api/FeedItemsMapper.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-struct RemoteFeedItem: Decodable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let image: URL
-}
-
 internal final class FeedItemsMapper {
     private struct Root: Decodable {
         let items: [RemoteFeedItem]
