+++ EssentialFeed/EssentialFeed/Feed Cache/FeedStore.swift
@@ -14,17 +14,3 @@ public protocol FeedStore {
     func deleteCachedFeed(completion: @escaping Deletioncompletion)
     func insert(_ items: [LocalFeedItem], with timestamp: Date, completion: @escaping Insertioncompletion)
 }
-
-public struct LocalFeedItem: Equatable {
-    public let id: UUID
-    public let description: String?
-    public let location: String?
-    public let imageURL: URL
-    
-    public init(id: UUID, description: String?, location: String?, imageURL: URL) {
-        self.id = id
-        self.description = description
-        self.location = location
-        self.imageURL = imageURL
-    }
-}
